% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{wkcompile_command}
\alias{wkcompile_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkcompile_command(compiler, cmd_json)
}
\arguments{
\item{compiler}{A compiler handle from [wkcompile_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
